feat: Add tests for assemble task
